+++ bingads/reporting/reporting_service_manager.py
@@ -45,6 +45,7 @@ class ReportingServiceManager:
         # make sure the working directory exists or create it.
         if not os.path.exists(self._working_directory):
             os.makedirs(self._working_directory)
+        self._suds_options = suds_options
 
     def download_file(self, download_parameters):
         """ Downloads the specified reporting to a local file.
@@ -88,7 +89,7 @@ class ReportingServiceManager:
             authorization_data=self._authorization_data,
             poll_interval_in_milliseconds=self._poll_interval_in_milliseconds,
             environment=self._environment,
-            location=self.service_client.service_url
+            **self.suds_options
         )
         return operation
 
@@ -127,3 +128,15 @@ class ReportingServiceManager:
     @working_directory.setter
     def working_directory(self, value):
         self._working_directory = value
+
+    @property
+    def suds_options(self):
+        """ suds option parameters
+
+        :return: dict
+        """
+        return self._suds_options
+
+    @suds_options.setter
+    def suds_options(self, value):
+        self._suds_options = value